qat_batch_cluster type casting to long

+++ utils/misc.py
@@ -289,7 +289,7 @@ def pcq_epoch(model, clustering_model, train_loader, criterion, optimizer, runti
     with tqdm(range(len(train_loader)), desc="Epoch {}".format(epoch), ncols=90) as t:
         for i, _ in enumerate(t):
             input, target, runtime_helper.batch_cluster = container.get_batch()
-            runtime_helper.qat_batch_cluster = torch.tensor(runtime_helper.batch_cluster, dtype=torch.int, device='cuda', requires_grad=False)
+            runtime_helper.qat_batch_cluster = torch.tensor(runtime_helper.batch_cluster, dtype=torch.int64, device='cuda', requires_grad=False)
             input, target = input.cuda(), target.cuda()
             output = model(input)
 
@@ -338,7 +338,7 @@ def pcq_validate(model, clustering_model, test_loader, criterion, runtime_helper
 
                 input, target, runtime_helper.batch_cluster = container.get_batch()
                 input, target = input.cuda(), target.cuda()
-                runtime_helper.qat_batch_cluster = torch.tensor(runtime_helper.batch_cluster, dtype=torch.int,
+                runtime_helper.qat_batch_cluster = torch.tensor(runtime_helper.batch_cluster, dtype=torch.int64,
                                                                 device='cuda', requires_grad=False)
                 output = model(input)
